requirements.txt
streamlit==1.49.1
torch
xgboost
lightgbm
catboost
pyhdf
python-dotenv
numpy<2
pandas==2.3.2
scikit-learn==1.7.2
plotly==6.3.0
rasterio>=1.3.0
Pillow>=9.0.0
earthengine-api>=0.1.384
geopy>=2.3.0
streamlit-folium>=0.18.0
folium>=0.17.0
